MAPL-3 - - Removed CocoaPods support - Added SPM manifest - Updated Example app - Fixed mistyping
