--- Example/Podfile
@@ -1,11 +1,0 @@
-use_frameworks!
-
-target 'MapLauncher_Example' do
-  pod 'MapLauncher', :path => '../'
-
-  target 'MapLauncher_Tests' do
-    inherit! :search_paths
-
-    
-  end
-end
