+++ MAPLExample/MAPLExample/MAPLExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  MAPLExampleApp.swift
+//  MAPLExample
+//
+//  Created by Dmytro Romanov on 30.05.2023.
+//
+
+import SwiftUI
+
+@main
+struct MAPLExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
