+++ Tests/MapLauncherTests/MapLauncherTests.swift
@@ -1,0 +1,11 @@
+import XCTest
+@testable import MapLauncher
+
+final class MapLauncherTests: XCTestCase {
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
+        XCTAssertEqual(MapLauncher().text, "Hello, World!")
+    }
+}
